Got values printing

+++ convolution/src/looptestarrayinput.c
@@ -3,29 +3,29 @@
 
 
 
-void testloops(double* input) {
+void testloops(float* input) {
     
 	
 	
-const double array[5][5] ={{  1,   2, 3, 4,  5},
+const float array[5][5] ={{  1,   2, 3, 4,  5},
 				{6,   7, 8, 9,    10},
 				{54,   5, 346, 64, 74},
 				{54, 54, 54,   125,  125},
 				{1, 125, 125,   125, 125}};
 	
 	
-	double tot = 0;
+	float tot = 0;
     for(int i=0; i<100; i++) {
         for (int j=0; j<100; j++) {
           for(int k=0; k<5; k++){
             for(int l=0; l<5; l++){
 				
-			  tot+= (double)input[k*5 + l] * array[k][l];
+			  tot+= (float)input[k*5 + l] * array[k][l];
 	        }
 	      }
 	    }
     }
-    fprintf(stdout, "%lf\n", tot);
+    fprintf(stdout, "%f\n", tot);
 }
 
 
@@ -33,9 +33,9 @@ const double array[5][5] ={{  1,   2, 3, 4,  5},
 
 
 int main(){
-	double input[25];
+	float input[25];
 	for (int i = 0; i < 25; ++i) {
-    	scanf("%lf", input + i);
+    	scanf("%f", input + i);
 	}
 
     testloops(input);
